+++ src/dip_coater/motor_driver/tmc2660_playground.py
@@ -6,7 +6,7 @@ import pytrinamic
 from pytrinamic.modules import Landungsbruecke
 
 from TMC_2209._TMC_2209_logger import Loglevel
-from dip_coater.motor.tmc2660 import MotorDriverTMC2660
+from dip_coater.motor_driver.tmc2660 import MotorDriverTMC2660
 from dip_coater.mechanical.mechanical_setup import MechanicalSetup
 
 pytrinamic.show_info()
