+++ src/test/test_tmc5160_hardware.py
@@ -4,7 +4,7 @@ from types import SimpleNamespace
 import pytest
 
 from dip_coater.mechanical.mechanical_setup import MechanicalSetup
-from dip_coater.motor.tmc5160 import MotorDriverTMC5160
+from dip_coater.motor_driver.tmc5160 import MotorDriverTMC5160
 
 
 class HardwareAppState:
